Add policy compliance metrics
When a policy is compliant, its status metric will be 0. When it is
noncompliant, the metric will be 1. The metric is labelled so that many
rich queries can be performed through prometheus.

Examples:

To find the number of clusters policy "foo" is propagated to:
`count(ocm_policy_status{type="propagated",name="foo"})`

To find the number of noncompliant root policies:
`count(ocm_policy_status{type="root"}==1)`

Refs:
 - https://github.com/open-cluster-management/backlog/issues/14727

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>