Add tests for the policy info metrics
Because the endpoint isn't directly exposed, it would take some extra
work to get it to a "regular" endpoint that the tests could connect to.
So, instead, this exec's into the pod, where it *can* curl the metrics
endpoint directly. The new util function could be useful for testing
other metrics in the future.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/14992

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>