Migrate main controller code to new operator-sdk
Previous imports of apis that had been copied into this repo are now
imports of external repos, where those apis/crds are actively updated.

The *Mappers (part of using predicates) were updated to to be MapFuncs,
as opposed to implementing a Map method. This was due to changes in the
controller runtime v0.7, see those release notes for more details:
 - https://github.com/kubernetes-sigs/controller-runtime/releases/tag/v0.7.0

The SetupWithManager method replaces the previous AddToManager/Add/add
functions, and is called for each controller in the main.go explicitly.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/15548

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>