Fix watches using new controller-runtime
In the course of the operator-sdk upgrade, the controller-runtime
changed some of the methods for creating a manager. In particular, users
now need to specify a "For" instead of just being able to define the
Watches, or the runtime will through an error. Unfortunately, the For
function does not allow for modifying the eventhandler.

The propagator and policyautomation controllers were impacted by this,
because they both used modified eventhandlers for their primary
resources. In the case of the policyautomation, the Mapper was able to
be integrated into the Reconcile method. But for the propagator, this
caused some reconcile loops to be missed, and tests were failing.

The workaround here is to define a For on the propagator manager, with
a predicate so that it is never triggered. Then a normal Watches is
defined with the previous Mapper, to work as it did previously.

Some additional RBAC permissions were also required.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/15548

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>