Fix all linting errors
As a team, we decided which linting rules we want to enforce:
https://github.com/open-cluster-management/backlog/issues/17089

Based on that discussion, the linters were configured and reenabled.
This commit also fixes all the linting errors.

One caveat is that `forcetypeassert` linting errors were disabled on the
offending lines since the calling code is from the controller runtime
which is validated from the Kubernetes API, so the type will always be
as expected.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>