Make the template config a local variable
Previous to this commit, the policy template configuration was a package
variable but a recent change made template process happen concurrently.
This makes modifications to the template configuration susceptible to
race conditions. The race conditions aren't currently hit because the
only change that is made is the lookup namespace restriction, which is
different for each policy but not for each placement decision, which is
where the concurrency happens.

This will make future changes easier that involve policy template
configuration changes per placement decision.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>