Adjust Makefile targets for create-or-update
Invocations of `kubectl create` will fail if the resource is already on
the cluster, but usually we want to ignore that. By piping `create` to
`apply` like this, we won't get an error, and the resource would be
updated if necessary.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 8bbe3fad4633463672a85608e7fb1475b9c47a68)