Prevent panic from unexpected label values
There were two places where the root-policy label was assumed to have
one `.` in it, since that's how the propagator sets the label. But if a
user created a Policy with that label, it doesn't necessarily have that
character, so the propagator could panic trying to access a value out
of the range of the slice.

There was another spot that used a similar split, but that one is safe,
for now.

Refs:
 - https://github.com/stolostron/backlog/issues/26592

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 2914c006d17e26c46cbaf42f74585971f58a5554)