Increase safety of decisionResults
Previously, the Identifier in this struct was a string, in the format
of `<namespace>/<name>`, and consumers would split that string to get
the pieces they want. This had the potential to cause slice index out of
range panics. Now the Identifier is a struct with the desired fields.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 08e0bc6430ea8557a4ff5de51d9084b9376a3a9d)