Refactor root policy status updates
In the event that a policy status constantly changes on hundreds of
managed clusters, the CPU and memory utilization would significantly
increase. This was due to status updates in replicated policies always
triggering the expensive code of generating all replicated policies and
updating them if necessary as well as calculate the root policy status
by looking at all the replicated policy statuses.

This new approach adds a new controller that handles just status updates
from replicated policies and then updates root policy status based on
the current state. The main policy propagator controller still gets
called for all other events.

As part of this, some refactoring was done to improve resource
utilization and performance. One example is instead of listing all
managed clusters to see if there is a managed cluster with the same name
as the policy namespace, a single get request is made. In the event of
3000 managed clusters, this significantly reduces memory utilization as
a list operation would have created 3000 copies from the
controller-runtime cache.

The retries logic around individual API calls was removed because almost
all of the failures can't be resolved with retries since they are gets
and lists from the controller-runtime cache. Update operations that fail
are likely due to the underlying object changing since it was retrieved
from the cache, so a retry on the whole reconcile is a better approach.
This is beneficial because a failed reconcile will get retried right
away now rather than wait for 5 minutes.

These changes in a local scale test reduced CPU and memory utilization
significantly. Memory and CPU utilization no longer grow linearly and
instead stabilize at low resource utilization.

Relates:
https://issues.redhat.com/browse/ACM-2543
https://issues.redhat.com/browse/ACM-2774

Signed-off-by: mprahl <mprahl@users.noreply.github.com>
(cherry picked from commit f59745dd8fe7bcdb9477fbbd9a429c631e0823bd)