Better handling of cluster-scoped template lookups
Cherry-pick to release-2.7 note: the new template-utils version caused
other dependencies to be updated, which required some other small code
changes.

Previously, there were some inconsistencies in how a hub template lookup
for a cluster-scoped resource worked. As a result, the propagator would
throw an error, and not be able to populate a real status for policies
that tried to use that kind of lookup.

Now, in general, using lookup in a hub template for a cluster-scoped
resource will fail with a template error that can be seen in the
propagated policy. The one current exception is that a lookup of the
ManagedCluster where the policy is replicated is allowed.

Refs:
 - https://issues.redhat.com/browse/ACM-5547

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 80bd4030d5de9a1921dda3ba1b391eb6ea9c2395)
(cherry picked from commit 278934edd4a69d751d97925e7d53b18e9576706e)