Refactor getting placements and decisions
Previously, the `common` package had separate functions to get the
decisions from a PlacementBinding depending on whether it was bound to
a Placement or a PlacementRule. But every caller would need to use both,
so that was unnecessary. Now, there is a unified `GetDecisions` func.

Similarly, the `propagator` had separate functions to find the policy
placements (not to be confused with Placements...), and the de-
duplicated logic was brought to the `getPolicyPlacementDecisions` func
more directly.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 4178f22748c05eafb42a4b2ed2fe72391a319d70)