Cache hub template queries
This leverages the new caching functionality in go-template-utils so
that API queries are not duplicated.

As a side-effect, this should cause less reconciles because when
starting a watch due to a hub template, it does not cause an initial
reconcile merely from the watch being created.

Relates:
https://issues.redhat.com/browse/ACM-7402
https://issues.redhat.com/browse/ACM-7398

Signed-off-by: mprahl <mprahl@users.noreply.github.com>
(cherry picked from commit df22cc22a27f95eb3a276e9697bb95849d5eb866)