Add partial indexes to account for null values
In Postgres 15+, NULLS NOT DISTINCT is supported so that a null value in
the same column is seen as the same value. This is not standard SQL and
Postgres 15 is likely too new to require. This is a workaround to
include partial indexes when some null values are provided.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>
(cherry picked from commit b8f123783a367c5325d6c9b072aada61b5da1a6b)