Add certificate and token auth and authz on the POST API endpoint
If a certificate is provided, the HTTP server will run over HTTPS which
enables certificate authentication.

The authorization requires a user to have patch access to the policy
status on the managed cluster namespace of the managed cluster that the
compliance event is for.

Relates:
https://issues.redhat.com/browse/ACM-6866

Signed-off-by: mprahl <mprahl@users.noreply.github.com>
(cherry picked from commit 857a4689e5144fa2bb1b90d13e5ee6094ad50fa3)