Send the correct content-type header with a large CSV
This seems to be a quirk in the Go HTTP server when sending chunks where
if there is more than one chunk, it seems to fallback to the
content-type originally set. Before this commit, it was
application/json.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>
(cherry picked from commit 83879f9ef039a766ba0f8227a6a7525e660f1293)