Fix index errors for large values
A Postgres unique constraint fails if the spec or message is too large. The
solution was to split out the spec to a separate table and include a
hash of the message that the unique constraint is on.

Relates:
https://issues.redhat.com/browse/ACM-10794

Signed-off-by: mprahl <mprahl@users.noreply.github.com>
(cherry picked from commit a7eda65dba49aaf74a0d865443f7e008e1ccf893)