Add flag to disable PlacementRule watches
Some deployments will not want to install the PlacementRule CRD, since
it is an older API which may eventually be removed. This adds an
argument to the propagator which will disable all watches on
PlacementRules, allowing the controllers to run in those environments.

In addition, if the flag is not set, the controller will check for the
PlacementRule CRD on the cluster, and if not found, disable those
watches (which would otherwise fail and prevent the controller from
working).

Refs:
 - https://issues.redhat.com/browse/ACM-2609

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 474fdc8eda15469a477eb9861461a94e86c35a0d)